Add more content to "Wild Ass Guess"
I need to split this into two posts.

The Value of a Wild-ass Guess is still good. It should describe the
order-of-magnitude game and the plumber story. It's too much to tie this
to the opportunity sizing workshop though.

Separately, I think the opportunity sizing workship piece is probably
worth-while. At least as a reference piece.